fix: move client scripts out of scripts directory
