+++ src/browser/client-bridge/index.js
@@ -11,7 +11,7 @@ exports.ClientBridge = ClientBridge;
 exports.build = async (browser, opts = {}) => {
     const needsCompatLib = opts.calibration && opts.calibration.needsCompatLib;
 
-    const scriptFileName = needsCompatLib ? "bundle.compat.js" : "bundle.js";
+    const scriptFileName = needsCompatLib ? "bundle.compat.js" : "bundle.native.js";
 
     if (bundlesCache[scriptFileName]) {
         return ClientBridge.create(browser, bundlesCache[scriptFileName]);
