+++ src/browser/client-scripts/.eslintrc.js
