+++ src/browser/client-scripts/ignore-areas.js
