+++ src/browser/client-scripts/lib.compat.js
