+++ src/browser/client-scripts/lib.native.js
