+++ src/browser/client-scripts/polyfills/getComputedStyle.js
