+++ src/browser/client-scripts/rect.js
