+++ src/browser/client-scripts/util.js
